+++ platform/jvm/capture/src/main/kotlin/io/bitdrift/capture/IBridge.kt
@@ -10,6 +10,8 @@ package io.bitdrift.capture
 import io.bitdrift.capture.error.IErrorReporter
 import io.bitdrift.capture.network.ICaptureNetwork
 import io.bitdrift.capture.providers.session.SessionCallback
+import io.bitdrift.capture.providers.session.SessionStrategyConfiguration
+import io.bitdrift.capture.providers.Field
 import io.bitdrift.capture.reports.IssueCallbackConfiguration
 
 internal interface IBridge {
@@ -36,5 +38,6 @@ internal interface IBridge {
         errorReporter: IErrorReporter,
         startInSleepMode: Boolean,
         issueCallbackConfiguration: IssueCallbackConfiguration?,
+        initialFields: Array<Field> = emptyArray(),
     ): Long
 }