+++ test/platform/swift/unit_integration/core/Logger+Tests.swift
@@ -17,6 +17,7 @@ extension Logger {
         sessionConfiguration: SessionConfiguration? = nil,
         dateProvider: DateProvider? = nil,
         fieldProviders: [FieldProvider] = [],
+        initialFields: Fields = [:],
         configuration: Configuration = .testConfiguration,
         loggerBridgingFactoryProvider: LoggerBridgingFactoryProvider = LoggerBridgingFactory()
     ) throws -> Logger
@@ -29,6 +30,7 @@ extension Logger {
                 sessionStrategy: sessionConfiguration.map(SessionStrategy.configuration) ?? sessionStrategy,
                 dateProvider: dateProvider,
                 fieldProviders: fieldProviders,
+                initialFields: initialFields,
                 storageProvider: MockStorageProvider(),
                 timeProvider: SystemTimeProvider(),
                 loggerBridgingFactoryProvider: loggerBridgingFactoryProvider